refactor: extract StartupInputConfigurator from App.xaml.cs
Move the mouse-in-pointer startup configuration out of the App
constructor into a dedicated StartupInputConfigurator helper. The
P/Invoke declarations, the OPENCLAW_DISABLE_MOUSE_IN_POINTER guard,
the env-var parsing and the configuration logic all move together; the
constructor now calls StartupInputConfigurator.Configure() in the same
spot, before InitializeComponent(), preserving startup ordering.

Pure extraction, no behavior change. Logging strings, guard order,
captured values and caught exceptions are identical to the original.
